Semi-broken link at (2 h3 with the same id)
At the 4.14 exercise there is a link to /en/part3/saving_data_to_mongo_db#other-operations which isn't the right one because there are 2 h3 with that id and it should link to the second one (the one where it teaches you how to update a note, not to create one). I don't know what change to propose because more files should be updated and I don't want to break something but it's something to look at.