Remove remark about addresses being for local dbs
Remove the remark in the parenthetical about the addresses being for local databases. If it was referring to something else, I couldn't find it in the document by looking ahead for `mongodb`, nor did I recognize what it was talking about because just prior the text said "we won't be doing a local database, we'll just keep using Mongo Atlas".